refactor(builder): extract build_loop_tools cluster to loop_tools.rs (dirge-4y4l stage 11b)
Stage 11b of the builder decomposition. Moves the LoopTool-registry
construction cluster out of builder/mod.rs into loop_tools.rs.

- loop_tools.rs (512 LOC) — build_loop_tools (~420-LOC god-function that
  assembles the Vec<Arc<dyn LoopTool>> the agent_loop dispatches against),
  wrap_mcp_tools (background-injected MCP tool wrapper), and the
  DynamicToolSearch handle struct. Re-exported via `pub use loop_tools::*`
  so external paths (builder::build_loop_tools / wrap_mcp_tools) are
  unchanged. Calls the shared `build_session_search_tool` via `super`.
- mod.rs (586 LOC, was 1068) — now build_agent_inner (~516) + hookify +
  build_session_search_tool + wiring. build_agent_inner calls
  build_loop_tools through the re-export. hookify stays (build_agent_inner
  -only). No import churn — build_agent_inner shares loop_tools' import set.

build_agent_inner extraction is stage 11c. Behavior-preserving.
-D warnings clean default + windows-default; 2374 tests green.

Refs dirge-w11r
